state0 comparison
saves set of readings in state dict, in the first extended window of readings, takes mode (most frequently occuring state) of readings and saves it as vacant_state.
identical transformation is applied to all other polled set of readings and compared to vacant_state. if they differ, then the state is not vacant. (but is it because of noise or a vehicle? who knows...)